[FEATURE][ML] Allow parsing different types of results from analytics process
